+++ utils.js
@@ -1,0 +1,8 @@
+export const pick = (obj, keys) => {
+  const result = {};
+  for (const key of keys) {
+    if (obj[key]) result[key] = obj[key];
+  }
+
+  return result;
+};